feat: expiration-timer moved to base-lobby-card, simplify look
